Modularity
Some modularity added
