+++ .gitignore
@@ -5,3 +5,4 @@
 *.xml
 *.pyc
 *.mp3
+.DS_Store
